requirements.txt
python-dotenv==0.20.0
requests==2.31.0
python-telegram-bot==13.13
pymongo==4.6.0